Move tests for utils.
